requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
dotenv==0.9.9
ffmpeg-python==0.2.0
Flask==3.1.2
flask-cors==6.0.1
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pymongo==4.15.4
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
schedule==1.2.2
sniffio==1.3.1
upload-post==0.1.2
urllib3==2.5.0
Werkzeug==3.1.3
